refactor proposal memo detail router
