+++ src/app/routers/proposal_memos.py
@@ -40,25 +40,3 @@ async def create_proposal_memo(
         idempotency_key=idempotency_key,
         correlation_id=correlation_id,
     )
-
-
-@router.get(
-    "/{proposal_id}/versions/{version_no}/memo",
-    response_model=ProposalMemoEnvelopeResponse,
-    summary="Get Proposal Memo",
-    description=(
-        "Returns the source-owned proposal memo from lotus-advise. Gateway does not recompute "
-        "suitability, readiness, supportability, archive refs, or memo sections locally."
-    ),
-)
-async def get_proposal_memo(
-    proposal_id: str = PROPOSAL_ID_PATH,
-    version_no: int = VERSION_NO_PATH,
-) -> ProposalMemoEnvelopeResponse:
-    service = proposal_service()
-    correlation_id = correlation_id_var.get()
-    return await service.get_proposal_memo(
-        proposal_id=proposal_id,
-        version_no=version_no,
-        correlation_id=correlation_id,
-    )
